Edits to PassengerIdDetail, and Service and IService
