--- ElevenFiftyFlights.Models/PassengerIdModels/PassengerIdDetail.cs
@@ -1,9 +1,0 @@
-namespace ElevenFiftyFlights.Models.User;
-
-public class PassengerIdDetail
-{
-    public int Id { get; set; }
-    public int UserId { get; set; }
-    public int ConfirmationNumber { get; set; }
-    public int FlightId { get; set; }
-}
