+++ ElevenFiftyFlights.Services/PassengerId/IPassengerIdService.cs
@@ -1,6 +1,9 @@
+using ElevenFiftyFlights.Models.PassengerId;
+
 namespace ElevenFiftyFlights.Services.PassengerId;
 
 public interface IPassengerIdService
 {
-    
+    Task<bool> BookingPassengerAsync(PassengerIdDetail model);
+    Task<PassengerIdDetail?> GetPassengerIdByIdAsync(int passengerId);
 }
