+++ ElevenFiftyFlights.Services/PassengerId/PassengerIdService.cs
@@ -1,6 +1,24 @@
+using ElevenFiftyFlights.Data.Entities;
+using ElevenFiftyFlights.Models.PassengerId;
+
 namespace ElevenFiftyFlights.Services.PassengerId;
 
 public class PassengerIdService : IPassengerIdService
 {
+    public async Task<bool> BookingPassengerAsync(PassengerIdDetail model)
+    {
+        PassengerIdEntity entity = new()
+        {
+            Id = model.Id,
+            UserId = model.UserId,
+            ConfirmationNumber = model.ConfirmationNumber,
+            FlightId = model.FlightId
+
+        };
+    }
     
+    public async Task<PassengerIdDetail?> GetPassengerIdByIdAsync(int passengerId);
+    {
+
+    }
 }
